Framework for what paths should be available.
