+++ assets/src/apps/authoring/components/Flowchart/FlowchartTopToolbar.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Button } from '../../../../components/misc/Button';
 
 interface FlowchartTopToolbarProps {
   onSwitchToAdvancedMode: () => void;
