+++ assets/src/apps/authoring/components/Flowchart/flowchart-selectors.ts
@@ -3,13 +3,7 @@ import {
   selectAllActivities,
 } from '../../../delivery/store/features/activities/slice';
 import { AuthoringRootState } from '../../store/rootReducer';
-import {
-  AllPaths,
-  AlwaysGoToPath,
-  correctPathTypes,
-  DestinationPath,
-  DestinationPaths,
-} from './flowchart-path-utils';
+import { AllPaths, AlwaysGoToPath, correctPathTypes, DestinationPaths } from './paths/path-types';
 
 // Returns a list of output paths for a screen.
 export const selectScreenPaths = (state: AuthoringRootState, screenId: number): AllPaths[] => {
